+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 0.1.1.0 (next)
 
 - Add EventCreatedOn Event attribute
+- Split Data module into specialized modules
+- Support more fields for GerritChange
 
 ## 0.1.0.0 (2021-05-30)